+++ lib/mason.dart
@@ -2,16 +2,17 @@
 /// generate files quickly and consistently.
 ///
 /// ```sh
-/// # activate mason
-/// pub global activate mason
+/// # Activate mason
+/// $ dart pub global activate mason
 ///
-/// # see usage
-/// mason --help
+/// # See usage
+/// $ mason --help
 /// ```
 library mason;
 
 export 'src/exception.dart' show MasonException;
-export 'src/generator.dart' show MasonGenerator, DirectoryGeneratorTarget;
+export 'src/generator.dart'
+    show MasonGenerator, DirectoryGeneratorTarget, FileConflictResolution;
 export 'src/logger.dart' show Logger;
 export 'src/mason_bundle.dart' show MasonBundle;
 export 'src/mason_yaml.dart' show GitPath;