+++ lib/src/command_runner.dart
@@ -23,9 +23,12 @@ class MasonCommandRunner extends CommandRunner<int> {
     addCommand(CacheCommand(logger: _logger));
     addCommand(BundleCommand(logger: _logger));
     addCommand(InitCommand(logger: _logger));
+    addCommand(InstallCommand(logger: _logger));
+    addCommand(ListCommand(logger: _logger));
     addCommand(GetCommand(logger: _logger));
     addCommand(MakeCommand(logger: _logger));
     addCommand(NewCommand(logger: _logger));
+    addCommand(UninstallCommand(logger: _logger));
   }
 
   final Logger _logger;
@@ -55,7 +58,7 @@ class MasonCommandRunner extends CommandRunner<int> {
   @override
   Future<int?> runCommand(ArgResults topLevelResults) async {
     if (topLevelResults['version'] == true) {
-      _logger.info('mason version: $packageVersion');
+      _logger.info('Mason $packageVersion');
       return ExitCode.success.code;
     }
     return super.runCommand(topLevelResults);