+++ lib/src/commands/commands.dart
@@ -2,5 +2,8 @@ export 'bundle.dart';
 export 'cache.dart';
 export 'get.dart';
 export 'init.dart';
+export 'install.dart';
+export 'list.dart';
 export 'make.dart';
 export 'new.dart';
+export 'uninstall.dart';