+++ lib/src/commands/make.dart
@@ -1,8 +1,10 @@
 import 'dart:convert';
 import 'dart:io';
+import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:io/ansi.dart';
 import 'package:io/io.dart';
+import 'package:path/path.dart' as p;
 import 'package:mason/mason.dart';
 import 'package:mason/src/generator.dart';
 
@@ -15,6 +17,7 @@ import '../command.dart';
 class MakeCommand extends MasonCommand {
   /// {@macro make_command}
   MakeCommand({Logger? logger}) : super(logger: logger) {
+    argParser.addOptions();
     try {
       for (final brick in bricks) {
         addSubcommand(_MakeCommand(brick, logger: logger));
@@ -46,11 +49,7 @@ class MakeCommand extends MasonCommand {
 
 class _MakeCommand extends MasonCommand {
   _MakeCommand(this._brick, {Logger? logger}) : super(logger: logger) {
-    argParser.addOption(
-      'json',
-      abbr: 'j',
-      help: 'Path to json file containing variables',
-    );
+    argParser.addOptions();
     for (final arg in _brick.vars) {
       argParser.addOption(arg);
     }
@@ -66,7 +65,17 @@ class _MakeCommand extends MasonCommand {
 
   @override
   Future<int> run() async {
-    final target = DirectoryGeneratorTarget(cwd, logger);
+    final outputDir = p.canonicalize(
+      p.join(cwd.path, results['output-dir'] as String),
+    );
+    final configPath = results['config-path'] as String?;
+    final fileConflictResolution =
+        (results['on-conflict'] as String).toFileConflictResolution();
+    final target = DirectoryGeneratorTarget(
+      Directory(outputDir),
+      logger,
+      fileConflictResolution,
+    );
 
     Function? generateDone;
     try {
@@ -75,10 +84,10 @@ class _MakeCommand extends MasonCommand {
       generateDone = logger.progress('Making ${generator.id}');
 
       try {
-        vars.addAll(await _decodeFile(results['json'] as String?));
+        vars.addAll(await _decodeFile(configPath));
       } on FormatException catch (error) {
         generateDone();
-        logger.err('${error}in ${results['json']}');
+        logger.err('${error}in $configPath');
         return ExitCode.usage.code;
       } on Exception catch (error) {
         generateDone();
@@ -90,15 +99,11 @@ class _MakeCommand extends MasonCommand {
         if (vars.containsKey(variable)) continue;
         final arg = results[variable] as String?;
         if (arg != null) {
-          vars.addAll(
-            <String, dynamic>{variable: _maybeDecode(arg)},
-          );
+          vars.addAll(<String, dynamic>{variable: _maybeDecode(arg)});
         } else {
-          vars.addAll(
-            <String, dynamic>{
-              variable: _maybeDecode(logger.prompt('$variable: '))
-            },
-          );
+          vars.addAll(<String, dynamic>{
+            variable: _maybeDecode(logger.prompt('$variable: '))
+          });
         }
       }
       final fileCount = await generator.generate(target, vars: vars);
@@ -108,7 +113,7 @@ class _MakeCommand extends MasonCommand {
           '${lightGreen.wrap('✓')} '
           'Generated $fileCount file(s):',
         )
-        ..flush(logger.success);
+        ..flush(logger.detail);
       return ExitCode.success.code;
     } on Exception catch (error) {
       generateDone?.call();
@@ -119,8 +124,8 @@ class _MakeCommand extends MasonCommand {
 
   Future<Map<String, dynamic>> _decodeFile(String? path) async {
     if (path == null) return <String, dynamic>{};
-    final jsonVarsContent = await File(path).readAsString();
-    return json.decode(jsonVarsContent) as Map<String, dynamic>;
+    final content = await File(path).readAsString();
+    return json.decode(content) as Map<String, dynamic>;
   }
 
   dynamic _maybeDecode(String value) {
@@ -128,6 +133,46 @@ class _MakeCommand extends MasonCommand {
       return json.decode(value);
     } catch (_) {
       return value;
+    }
+  }
+}
+
+extension on ArgParser {
+  void addOptions() {
+    addOption(
+      'config-path',
+      abbr: 'c',
+      help: 'Path to config json file containing variables.',
+    );
+    addOption(
+      'output-dir',
+      abbr: 'o',
+      help: 'Directory where to output the generated code.',
+      defaultsTo: '.',
+    );
+    addOption(
+      'on-conflict',
+      allowed: ['prompt', 'overwrite', 'skip'],
+      defaultsTo: 'prompt',
+      allowedHelp: {
+        'prompt': 'Always prompt the user for each file conflict.',
+        'overwrite': 'Always overwrite conflicting files.',
+        'skip': 'Always skip conflicting files.'
+      },
+      help: 'File conflict resolution strategy.',
+    );
+  }
+}
+
+extension on String {
+  FileConflictResolution toFileConflictResolution() {
+    switch (this) {
+      case 'skip':
+        return FileConflictResolution.skip;
+      case 'overwrite':
+        return FileConflictResolution.overwrite;
+      default:
+        return FileConflictResolution.prompt;
     }
   }
 }