+++ test/commands/make_test.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: no_adjacent_strings_in_list
 import 'dart:convert';
 import 'dart:io';
 
@@ -29,6 +30,8 @@ void main() {
     path: ../../../bricks/documentation
   greeting:
     path: ../../../bricks/greeting
+  hello_world:
+    path: ../bricks/hello_world
   todos:
     path: ../../../bricks/todos
   widget:
@@ -43,6 +46,9 @@ void main() {
       );
       final greetingPath = path.canonicalize(
         path.join(Directory.current.path, bricksPath, 'greeting'),
+      );
+      final helloWorldPath = path.canonicalize(
+        path.join(Directory.current.path, bricksPath, 'hello_world'),
       );
       final todosPath = path.canonicalize(
         path.join(Directory.current.path, bricksPath, 'todos'),
@@ -59,11 +65,14 @@ void main() {
               docPath,
           '''greeting_81a4ec348561cdd721c3bb79b3d6dc14738bf17f02e18810dad2a6d88732e298''':
               greetingPath,
+          '''hello_world_6bb398ce3cd00f53423469ca6f0fbe8e6ee49a7351c5bf45de7faff464d31333''':
+              helloWorldPath,
           '''todos_6d110323da1d9f3a3ae2ecc6feae02edef8af68ca329601f33ee29e725f1f740''':
               todosPath,
           '''widget_02426be7ece33230d574cb7a76eb7a9a595a79cbf53a1b1c8f2f1de78dfbe23f''':
               widgetPath,
         }));
+      printLogs = [];
       logger = MockLogger();
       when(() => logger.progress(any())).thenReturn(([String? _]) {});
       commandRunner = MasonCommandRunner(logger: logger);
@@ -72,6 +81,65 @@ void main() {
     tearDown(() {
       Directory.current = cwd;
     });
+
+    test('--help shows correct help information', overridePrint(() async {
+      const expectedPrintLogs = <String>[
+        'Generate code using an existing brick template.\n'
+            '\n'
+            'Usage: mason make <subcommand> [arguments]\n'
+            '-h, --help                      Print this usage information.\n'
+            '''-c, --config-path               Path to config json file containing variables.\n'''
+            '''-o, --output-dir                Directory where to output the generated code.\n'''
+            '                                (defaults to ".")\n'
+            '''    --on-conflict               File conflict resolution strategy.\n'''
+            '\n'
+            '''          [overwrite]           Always overwrite conflicting files.\n'''
+            '''          [prompt] (default)    Always prompt the user for each file conflict.\n'''
+            '          [skip]                Always skip conflicting files.\n'
+            '\n'
+            'Available subcommands:\n'
+            '  app_icon        Create an app_icon file from a URL\n'
+            '  documentation   Create Documentation Markdown Files\n'
+            '  greeting        A Simple Greeting Template\n'
+            '  hello_world     A Simple Hello World Template\n'
+            '  todos           A Todos Template\n'
+            '  widget          Create a Simple Flutter Widget\n'
+            '\n'
+            'Run "mason help" to see global options.'
+      ];
+      final result = await commandRunner.run(['make', '-h']);
+      expect(result, equals(ExitCode.success.code));
+      expect(printLogs, equals(expectedPrintLogs));
+    }));
+
+    test('<subcommand> --help shows correct help information',
+        overridePrint(() async {
+      const expectedPrintLogs = <String>[
+        'A Simple Greeting Template\n'
+            '\n'
+            'Usage: mason make greeting [arguments]\n'
+            '-h, --help                      Print this usage information.\n'
+            '''-c, --config-path               Path to config json file containing variables.\n'''
+            '''-o, --output-dir                Directory where to output the generated code.\n'''
+            '                                (defaults to ".")\n'
+            '''    --on-conflict               File conflict resolution strategy.\n'''
+            '\n'
+            '''          [overwrite]           Always overwrite conflicting files.\n'''
+            '''          [prompt] (default)    Always prompt the user for each file conflict.\n'''
+            '          [skip]                Always skip conflicting files.\n'
+            '\n'
+            '    --name                      \n'
+            '\n'
+            'Run "mason help" to see global options.'
+      ];
+      final testDir = Directory(
+        path.join(Directory.current.path, 'greeting'),
+      )..createSync(recursive: true);
+      Directory.current = testDir.path;
+      final result = await commandRunner.run(['make', 'greeting', '--help']);
+      expect(result, equals(ExitCode.success.code));
+      expect(printLogs, equals(expectedPrintLogs));
+    }));
 
     test('exits with code 64 when brick does not exist', () async {
       final result = await commandRunner.run(['make', 'garbage']);
@@ -91,15 +159,17 @@ void main() {
       ).called(1);
     });
 
-    test('exits with code 64 when mason.yaml does not exist', () async {
-      File(path.join(Directory.current.path, 'mason.yaml'))
-          .deleteSync(recursive: true);
+    test('exits with code 64 when local mason.yaml does not exist', () async {
+      try {
+        File(path.join(Directory.current.path, 'mason.yaml'))
+            .deleteSync(recursive: true);
+      } catch (_) {}
       commandRunner = MasonCommandRunner(logger: logger);
       final result = await commandRunner.run(['make', 'garbage']);
       expect(result, equals(ExitCode.usage.code));
       verify(
         () => logger.err(
-          'Could not find mason.yaml.\nDid you forget to run mason init?',
+          'Could not find a subcommand named "garbage" for "mason make".',
         ),
       ).called(1);
     });
@@ -114,7 +184,7 @@ void main() {
       final result = await commandRunner.run([
         'make',
         'todos',
-        '--json',
+        '--config-path',
         'todos.json',
       ]);
       expect(result, equals(ExitCode.usage.code));
@@ -198,6 +268,28 @@ in todos.json''',
       expect(directoriesDeepEqual(actual, expected), isTrue);
     });
 
+    test('generates hello_world', () async {
+      final testDir = Directory(
+        path.join(Directory.current.path, 'hello_world'),
+      )..createSync(recursive: true);
+      Directory.current = testDir.path;
+      final result = await commandRunner.run([
+        'make',
+        'hello_world',
+        '--name',
+        'dash',
+      ]);
+      expect(result, equals(ExitCode.success.code));
+
+      final actual = Directory(
+        path.join(testFixturesPath(cwd, suffix: '.make'), 'hello_world'),
+      );
+      final expected = Directory(
+        path.join(testFixturesPath(cwd, suffix: 'make'), 'hello_world'),
+      );
+      expect(directoriesDeepEqual(actual, expected), isTrue);
+    });
+
     test('generates todos', () async {
       final testDir = Directory(
         path.join(Directory.current.path, 'todos'),
@@ -215,7 +307,7 @@ in todos.json''',
       final result = await commandRunner.run([
         'make',
         'todos',
-        '--json',
+        '-c',
         'todos.json',
       ]);
       expect(result, equals(ExitCode.success.code));
@@ -249,6 +341,108 @@ in todos.json''',
         path.join(testFixturesPath(cwd, suffix: 'make'), 'widget'),
       );
       expect(directoriesDeepEqual(actual, expected), isTrue);
+    });
+
+    test('generates greeting with custom output directory', () async {
+      final result = await commandRunner.run(
+        [
+          'make',
+          'greeting',
+          '--name',
+          'test-name',
+          '-o',
+          path.join('output_dir', 'dir')
+        ],
+      );
+      expect(result, equals(ExitCode.success.code));
+
+      final actual = Directory(
+        path.join(testFixturesPath(cwd, suffix: '.make'), 'output_dir', 'dir'),
+      );
+      final expected = Directory(
+        path.join(testFixturesPath(cwd, suffix: 'make'), 'output_dir', 'dir'),
+      );
+      expect(directoriesDeepEqual(actual, expected), isTrue);
+    });
+
+    test('generates greeting and skips conflicts', () async {
+      final testDir = Directory(
+        path.join(Directory.current.path, 'greeting-skip'),
+      )..createSync(recursive: true);
+      Directory.current = testDir.path;
+      var result = await commandRunner.run([
+        'make',
+        'greeting',
+        '--name',
+        'test-name',
+      ]);
+      expect(result, equals(ExitCode.success.code));
+
+      final fileA = File(
+        path.join(Directory.current.path, 'GREETINGS.md'),
+      );
+      expect(fileA.readAsStringSync(), contains('Hi test-name!'));
+      verify(
+        () => logger.delayed(any(that: contains('(new)'))),
+      ).called(1);
+
+      result = await commandRunner.run([
+        'make',
+        'greeting',
+        '--name',
+        'test-name2',
+        '--on-conflict',
+        'skip',
+      ]);
+
+      expect(result, equals(ExitCode.success.code));
+      final fileB = File(
+        path.join(Directory.current.path, 'GREETINGS.md'),
+      );
+      expect(fileB.readAsStringSync(), contains('Hi test-name!'));
+      verify(
+        () => logger.delayed(any(that: contains('(skip)'))),
+      ).called(1);
+    });
+
+    test('generates greeting and overwrites conflicts', () async {
+      final testDir = Directory(
+        path.join(Directory.current.path, 'greeting-overwrite'),
+      )..createSync(recursive: true);
+      Directory.current = testDir.path;
+      var result = await commandRunner.run([
+        'make',
+        'greeting',
+        '--name',
+        'test-name',
+      ]);
+      expect(result, equals(ExitCode.success.code));
+
+      final fileA = File(
+        path.join(Directory.current.path, 'GREETINGS.md'),
+      );
+      expect(fileA.readAsStringSync(), contains('Hi test-name!'));
+      verify(
+        () => logger.delayed(any(that: contains('(new)'))),
+      ).called(1);
+
+      result = await commandRunner.run([
+        'make',
+        'greeting',
+        '--name',
+        'test-name2',
+        '--on-conflict',
+        'overwrite',
+      ]);
+
+      expect(result, equals(ExitCode.success.code));
+      final fileB = File(
+        path.join(Directory.current.path, 'GREETINGS.md'),
+      );
+      expect(fileB.readAsStringSync(), contains('Hi test-name2!'));
+      verify(
+        () => logger.delayed(any(that: contains('(new)'))),
+      ).called(1);
     });
   });
 }